feat(macos): compile full service + model + theme layer on macOS
The Mac destination's INCLUDED_SOURCE_FILE_NAMES now covers the
entire cross-platform graph that TigerDuck's UI sits on top of:

- App/                AppConstants, AppDefaults, AppState,
                      BrowserPreference, LanguageManager,
                      WidgetURLRouter
- Models/Domain/      every file (now incl. CourseTimeSlot)
- Models/SwiftData/   SDAnnouncement, SDAssignment, SDCalendarEvent,
                      SDCourse
- Services/Auth/      AuthService, KeychainManager, SecureStore,
                      SSOWebAuthSession
- Services/Core/      DataCache, HTMLParser, NameAbbrService,
                      NetworkMonitor
- Services/Logging/   AppLogger
- Services/Migrations/  all four migrations
- Services/Push/      PushServerConfig (new — see below)
- Services/API/       Bulletin, Calendar, Library, Moodle, NTUST
                      (every Codable + HTTP client; NTUSTSessionManager
                      already cross-platform via 6166539)
- Theme/              every file incl. View+SurfaceStyle (now uses
                      Color.secondarySystemGroupedBackgroundCompat)
- Bridge/             AppServiceBridge, ModelAdapters
- Clock/              AppClock + state + override + TimezoneObserver
- Extensions/         every file except UINavigationController+Swipeback
- SharedUI/           NTUSTProtectedAccessState
- Shared/Theme/       VisualPreset

Mac build: ** BUILD SUCCEEDED **
iOS build: ** BUILD SUCCEEDED ** (no behaviour change)

PushServerConfig extraction
---------------------------

BulletinAPIClient.init default-args used PushCoordinator.resolveServerURL()
and resolveSharedSecret(). PushCoordinator imports ActivityKit + UIKit,
so anything that referenced those resolvers transitively blocked Mac
compilation of the bulletin path.

Moved the URL/secret/Secrets.plist resolvers into a new
`Services/Push/PushServerConfig.swift` enum (cross-platform). The
methods are byte-for-byte the same as the previous
PushCoordinator.resolveXxx surface — only the location and references
changed. assertEnvConsistency stays on PushCoordinator because it
checks PushAPNsEnv, which is iOS-only.

Updated call sites: PushCoordinator (init + assertEnvConsistency),
PushAPIClient (default arg), BulletinAPIClient (default arg),
BulletinTaxonomyStore, BulletinDetailView, BulletinSubscriptionsStore,
BulletinsViewModel.

What's still excluded from the Mac compile
------------------------------------------

Every Feature view, the LiveActivity bundle, Services/Push/* (minus
PushServerConfig), Services/Watch/*, Push delegate, NTUSTLoginSheetHost,
UINavigationController+Swipeback, PasswordField, SourceCodePickerView,
plus the iOS-only haptics / pasteboard / openSettingsURL sprinkles in
Settings + Bulletin + CourseDetail + Home + TimeSlider views. These
will move over feature-by-feature behind a real Mac UI.
