+++ swift/TigerDuck/Services/API/Bulletin/BulletinAPIClient.swift
@@ -14,7 +14,7 @@ final class BulletinAPIClient: Sendable {
     private let logger = Logger(subsystem: "org.ntust.app.TigerDuck", category: "Bulletin.API")
 
     init(
-        baseURL: URL = PushCoordinator.resolveServerURL(),
+        baseURL: URL = PushServerConfig.resolveServerURL(),
         session: URLSession? = nil,
         sharedSecret: String? = nil
     ) {
